# Create multicore flambda opam switch