--- a/bytecomp/bytegen.ml
+++ b/bytecomp/bytegen.ml
@@ -341,8 +341,8 @@ let comp_primitive p sz args =
   | Psetglobal id -> Ksetglobal id
   | Pintcomp cmp -> Kintcomp cmp
   | Pmakeblock(tag, _mut, _) -> Kmakeblock(List.length args, tag)
-  | Pfield(n, _ptr, Immutable) -> Kgetfield n
-  | Pfield(n, _ptr, Mutable) -> Kgetmutablefield n
+  | Pfield(n, _ptr, Immutable, _) -> Kgetfield n
+  | Pfield(n, _ptr, Mutable, _) -> Kgetmutablefield n
   | Pfield_computed -> Kgetvectitem
   | Psetfield(n, _ptr, _init) -> Ksetfield n
   | Psetfield_computed(_ptr, _init) -> Ksetvectitem