--- a/bytecomp/symtable.ml
+++ b/bytecomp/symtable.ml
@@ -157,7 +157,8 @@ let init () =
       let cst = Const_block(Obj.object_tag,
                             [Const_base(Const_string (name, None));
                              Const_base(Const_int (-i-1))
-                            ])
+                            ],
+                            Tag_none)
       in
       literal_table := (c, cst) :: !literal_table)
     Runtimedef.builtin_exceptions;
@@ -223,7 +224,7 @@ let rec transl_const = function
   | Const_base(Const_nativeint i) -> Obj.repr i
   | Const_pointer i -> Obj.repr i
   | Const_immstring s -> Obj.repr s
-  | Const_block(tag, fields) ->
+  | Const_block(tag, fields, _) ->
       let block = Obj.new_block tag (List.length fields) in
       let pos = ref 0 in
       List.iter