--- a/lambda/lambda.ml
+++ b/lambda/lambda.ml
@@ -221,11 +221,14 @@ let equal_value_kind x y =
   | Pintval, Pintval -> true
   | (Pgenval | Pfloatval | Pboxedintval _ | Pintval), _ -> false
 
+type tag_info =
+  | Tag_none
+  | Tag_record
 
 type structured_constant =
     Const_base of constant
   | Const_pointer of int
-  | Const_block of int * structured_constant list
+  | Const_block of int * structured_constant list * tag_info
   | Const_float_array of string list
   | Const_immstring of string