--- a/lambda/translobj.ml
+++ b/lambda/translobj.ml
@@ -26,7 +26,7 @@ let consts : (structured_constant, Ident.t) Hashtbl.t = Hashtbl.create 17
 
 let share c =
   match c with
-    Const_block (_n, l) when l <> [] ->
+    Const_block (_n, l, _) when l <> [] ->
       begin try
         Lvar (Hashtbl.find consts c)
       with Not_found ->