This PR
* Updates `ocaml-variants.opam` to include `flambda`
* Migrates my changes from `gc_parallel_minor` branch (needed for `ocaml2mcore` compiler).

See the file `ocaml-variants.opam` for how to set up a multicore opam switch with `flambda` enabled. The switch works to build Miking, but ironically the parallel programming feature does not work because it's not possible to install the `domainslib` package on this switch (yet).

More issues:
* The instruction (from the ifle `ocaml-variants.opam`) on how to update the opam switch:
```
opam install --assume-built ocaml-variants
```
freezes on my machine.
* When using this branch to compile `ocaml2mcore`, then I get an error on some programs. E.g.
```ocaml
let s = String.length "hello"
```
gives (when running `ocaml2mcore prog.ml`)
```
File "_none_", line 1, characters 22-29:
Error: This expression has type string/1
       but an expression was expected of type string/2
       string is abstract because no corresponding cmi file was found in path.
       File "_none_", line 1:
         Definition of type string/1
```
For now, `ocaml2mcore` can be built using the `gc_parallel_minor` branch.