@lingmar 2021-05-10
Some micro-benchmark results:
Before:
```
» ./run factorial 1000000 124
3. Miking compiler:      0.333597s  0.248092s
5. OCaml native:         0.312075s  0.249655s

./run factorial 2000
1. Boot interpreter:     0.112569s  0.109617s
2. Miking interpreter:   1.389382s  1.354209s
3. Miking compiler:      0.096622s  0.007602s
4. OCaml byte code       0.009635s  0.010307s
5. OCaml native:         0.068185s  0.006300s

» ./run tree 10000 2
1. Boot interpreter:     0.042248s  0.043437s
3. Miking compiler:      0.098945s  0.007692s

./run tree 100000000 12
3. Miking compiler:      1.024938s  0.940995s
```

After:
```
» ./run factorial 1000000 124
3. Miking compiler:      0.291776s  0.203526s
5. OCaml native:         0.304718s  0.247461s

» ./run factorial 2000
1. Boot interpreter:     0.113725s  0.116406s
2. Miking interpreter:   1.058818s  1.049596s
3. Miking compiler:      0.093765s  0.007303s
4. OCaml byte code       0.010231s  0.010362s
5. OCaml native:         0.069034s  0.006261s

» ./run tree 10000 2
1. Boot interpreter:     0.040883s  0.042318s
3. Miking compiler:      0.092587s  0.007180s

./run tree 100000000 12
3. Miking compiler:      0.943615s  0.855843s
```